producer/requirements.txt
base58==2.1.1
certifi==2021.10.8
charset-normalizer==2.0.11
cytoolz==0.11.2
ecdsa==0.17.0
eth-hash==0.3.2
eth-keys==0.4.0
eth-typing==3.0.0
eth-utils==2.0.0
idna==3.3
kafka-python==2.0.2
more-itertools==8.12.0
py-bip39-bindings==0.1.8
py-ed25519-bindings==1.0.1
py-sr25519-bindings==0.1.4
pycryptodome==3.14.1
requests==2.27.1
scalecodec==1.0.32
six==1.16.0
substrate-interface==1.2.0
toolz==0.11.2
urllib3==1.26.8
websocket-client==1.2.3
xxhash==2.0.2
